+++ frontend/src/pages/broker/queues/BrokerQueues.vue
@@ -4,7 +4,10 @@ import { useUserSettings } from '@/composables/userSettingsComposable'
 import { FilterMatchMode } from '@primevue/core/api'
 import Column from 'primevue/column'
 import DataTable, { type DataTableSortEvent } from 'primevue/datatable'
-import { computed, ref } from 'vue'
+import IconField from 'primevue/iconfield'
+import InputIcon from 'primevue/inputicon'
+import InputText from 'primevue/inputtext'
+import { computed, ref, watchEffect } from 'vue'
 import { useRouter } from 'vue-router'
 
 const router = useRouter()
@@ -23,8 +26,17 @@ const selectQueue = (data: any) => {
   })
 }
 
+const search = ref(settings.queueSearch)
+
+watchEffect(() => {
+  updateSettings({
+    ...settings,
+    queueSearch: search.value,
+  })
+})
+
 const filters = ref({
-  queueName: { value: null, matchMode: FilterMatchMode.CONTAINS },
+  queueName: { value: search, matchMode: FilterMatchMode.CONTAINS },
 })
 
 const onSort = (e: DataTableSortEvent) => {
@@ -54,30 +66,24 @@ const onSort = (e: DataTableSortEvent) => {
       @row-select="(e) => selectQueue(e.data)"
       @sort="onSort"
     >
-      <Column sortable field="queueName" header="Name" class="overflow-hidden overflow-ellipsis">
-        <template #body="{ data }">
-          {{ data.queueName }}
-        </template>
-        <template #filter="{ filterModel, filterCallback }">
-          <InputText
-            v-model="filterModel.value"
-            type="text"
-            @input="filterCallback()"
-            placeholder="Search by queue name"
-          />
+      <Column field="queueName" class="overflow-hidden overflow-ellipsis py-0">
+        <template #header>
+          <div class="flex w-full items-center">
+            <b>Name</b>
+            <IconField class="ms-1 grow">
+              <InputIcon class="pi pi-search" v-if="!search" />
+              <InputIcon class="pi pi-times cursor-pointer" v-else @click="search = ''" />
+              <InputText
+                placeholder="Search anything..."
+                class="w-full border-0 shadow-none dark:bg-surface-900"
+                v-model="search"
+                ref="searchInputText"
+              />
+            </IconField>
+          </div>
         </template>
       </Column>
-      <Column sortable field="ready" header="Ready" class="w-[0]"> </Column>
-      <Column sortable field="errored" header="Errored" class="w-[0]"></Column>
-      <Column
-        sortable
-        field="deadLettered"
-        header="Dead Lettered"
-        class="w-[0] whitespace-nowrap"
-      ></Column>
-      <Column sortable field="scheduled" header="Scheduled" class="w-[0]"></Column>
-      <Column sortable field="locked" header="Locked" class="w-[0]"></Column>
-      <Column sortable field="queueAutoDelete" header="Auto Delete" class="w-[0] whitespace-nowrap">
+      <Column field="queueAutoDelete" header="Auto Delete" class="w-[0] whitespace-nowrap">
         <template #body="{ data }">
           <div v-show="data['queueAutoDelete']" class="text-nowrap">
             {{ data['queueAutoDelete'] / 60 }}m
@@ -85,6 +91,42 @@ const onSort = (e: DataTableSortEvent) => {
           <div v-show="!data['queueAutoDelete']">-</div>
         </template>
       </Column>
+      <Column
+        field="queueMaxDeliveryCount"
+        header="Max Delivery"
+        class="w-0 whitespace-nowrap"
+      ></Column>
+      <Column
+        sortable
+        field="ready"
+        header="Ready"
+        class="w-[0] bg-surface-200/25 dark:bg-surface-800/20"
+      >
+      </Column>
+      <Column
+        sortable
+        field="errored"
+        header="Errored"
+        class="w-[0] bg-surface-200/25 dark:bg-surface-800/20"
+      ></Column>
+      <Column
+        sortable
+        field="deadLettered"
+        header="Dead Lettered"
+        class="w-[0] whitespace-nowrap bg-surface-200/25 dark:bg-surface-800/20"
+      ></Column>
+      <Column
+        sortable
+        field="scheduled"
+        header="Scheduled"
+        class="w-[0] bg-surface-200/25 dark:bg-surface-800/20"
+      ></Column>
+      <Column
+        sortable
+        field="locked"
+        header="Locked"
+        class="w-[0] bg-surface-200/25 dark:bg-surface-800/20"
+      ></Column>
     </DataTable>
   </div>
 </template>