+++ frontend/src/pages/broker/topics/BrokerTopics.vue
@@ -2,26 +2,28 @@
 import { useSubscriptionsQuery } from '@/api/subscriptions/subscriptionsQuery'
 import { useUserSettings } from '@/composables/userSettingsComposable'
 import { FilterMatchMode } from '@primevue/core/api'
-import { computed, ref } from 'vue'
+import { computed, ref, watchEffect } from 'vue'
 
-const { settings } = useUserSettings()
+const { settings, updateSettings } = useUserSettings()
 
 const { data: subscriptions } = useSubscriptionsQuery(computed(() => settings.refetchInterval))
 
+const search = ref(settings.topicSearch)
+
+watchEffect(() => {
+  updateSettings({
+    ...settings,
+    topicSearch: search.value,
+  })
+})
+
 const filters = ref({
-  topicName: { value: null, matchMode: FilterMatchMode.CONTAINS },
+  topicName: { value: search, matchMode: FilterMatchMode.CONTAINS },
 })
 </script>
 
 <template>
   <div class="flex overflow-auto">
-    <!-- 
-        :sort-field="settings.sortField"
-        :sort-order="settings.sortOrder" 
-        @row-select="(e) => selectQueue(e.data)"
-        @sort="onSort"
-        selection-mode="single"
-      -->
     <DataTable
       scrollable
       scroll-height="flex"
@@ -32,19 +34,21 @@ const filters = ref({
       v-model:filters="filters"
       filter-display="menu"
     >
-      <Column
-        sortable
-        field="topicName"
-        header="Topic Name"
-        class="overflow-hidden overflow-ellipsis"
-      >
-        <template #filter="{ filterModel, filterCallback }">
-          <InputText
-            v-model="filterModel.value"
-            type="text"
-            @input="filterCallback()"
-            placeholder="Search by topic name"
-          />
+      <Column field="topicName" class="overflow-hidden overflow-ellipsis py-0">
+        <template #header>
+          <div class="flex w-full items-center">
+            <b>Name</b>
+            <IconField class="ms-1 grow">
+              <InputIcon class="pi pi-search" v-if="!search" />
+              <InputIcon class="pi pi-times cursor-pointer" v-else @click="search = ''" />
+              <InputText
+                placeholder="Search anything..."
+                class="w-full border-0 shadow-none dark:bg-surface-900"
+                v-model="search"
+                ref="searchInputText"
+              />
+            </IconField>
+          </div>
         </template>
       </Column>
       <Column sortable field="routingKey" header="Routing Key" class="w-[0] whitespace-nowrap">