+++ CHANGELOG.md
@@ -2,6 +2,22 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [v1.3.1] - 2025-01-__
+
+### Added
+
+- Queue max delivery count
+
+### Changed
+
+- Updated to MassTransit v8.3.5
+- Search filter better UX
+- Table row paddings for better preview
+
+### Fixed
+
+- Selected tab color in dark mode
+
 ## [v1.3.0] - 2025-01-09
 
 ### Added